\name{trim_network}
\alias{trim_network}

\title{
Removes any columns or rows with only zeros. \cr
}
%~
\description{
Removes any columns or rows with only zeros.
}
%~
\usage{
trim_network(my_netowrk)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{my_netowrk}{incidence matrix representing the network}
}
%~
\details{
%% ~
}
\value{
matrix
%%
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Fernando Pedraza
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
%%  ~~further notes~~
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~incidence matrix}% use one of  RShowDoc("KEYWORDS")


